+++ Release_Notes.md
@@ -8,6 +8,7 @@
 ### Improvements
 - Split Web.Angular into Web.Api and Web.Angular, removed all .Net code/items
 - Authenticate/Logoff now returns both the authentication response and the client choices for improved performance.
+- Improved AbstractAccountController.ipAddress will use headers if the IP Address is not using the first method
 
 ### Bug Fixes
 - Fixed AuthorizeAttribute.OnAuthorization from getting a null result when getting MAccountProfile